FWAppInfo中のCameraInfoとDragInfoを整備
FWAppに直接書かれていた視点操作や剛体ドラッグ機能を各クラスへ移転しました。
これでFWAppのイベントハンドラをオーバライドすることでマウスボタン・キー割り当てを変更できるようになりました。

＊若干仕様レベルで変更あります：
・　注視点変更（中ボタンドラッグ）のゲインがやや変わりました
・　剛体ドラッグはシーン別に保持していましたが、過剰だと思ったので1つにしました。SwitchSceneでバグるかもしれません。

＊その他小さな修正

git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4767 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ include/Base/TMatrix.h
@@ -759,7 +759,7 @@ public:
 	///	o—Í
 	void print(std::ostream& os, char* sep="( )") const {
 //		os << "sz:" << height() << "," << width() << std::endl;
-		int w = os.width();
+		int w = (int)os.width();
 		os.width(0);
 		for(size_t i=0; i<height(); ++i){
 			for(int j=0; j<w; ++j){
